delete list=>list-dlt
This was intended to be a lossless function for converting a list of
pairs into a delta dict, but it's not actually lossless, because it
doesn't know the exact indices the items had in the original list. It's
still somewhat interesting and useful, but not enough so to say that
it's a necessary component of our list <=> delta-dict conversion theory,
which we barely mention in the paper. For now we'll delete it and focus
our list <=> delta-dict conversion theory on the ordinary list=>dlt
function.