complete list <=> dd conversion metatheory
Adds a new theorem which states that any key in l must also be in
list=>dlt l, and proves the inversion theorem. I think this should
suffice for the list <=> dlt metatheory. We don't prove every
conceivable theorem about list=>dlt, an operation whose metatheory is
somewhat awkward on account of it being lossy/non-injective. But I
think we've proven enough so to ensure that the list=>dlt function
behaves as expected.