resolve difference TODO
Now that all the private stuff is factored into a separate file, there's
no evidence of difference in the main file. The main file, and its
public definitions and theorems, is completely ignorant of
subtraction/difference, and agnostic as to how it would be implemented,
so this discussion is very much an implementation detail. I went ahead
and added a brief discussion about this into Nat.agda, but it's
essentially unnecessary.